requirements.txt
annotated-types==0.7.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.6
packaging==25.0
playwright==1.52.0
preshed==3.0.9
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
